+++ NEWS.md
@@ -9,6 +9,7 @@
 - `get_pipelines()`, `get_workflows()` and `get_jobs()` are now formatted as class `circle_builds`, `circle_collection()` and have a somewhat pretty print output
 - `*_checkout_key()`: Optimize printer, catch errors, add info messages, add test
 - make `get_pipelines()`, `get_workflows()` and `get_jobs()` work with API v2
+- rename `list_artifacts()` -> `get_build_artifacts()`
 
 ## Bugfixes